Move select component
