+++ packages/harmony/src/components/input/index.ts
@@ -3,4 +3,4 @@ export * from './TextInput'
 export * from './useFocusState'
 export * from './PasswordInput'
 export * from './TokenAmountInput'
-export * from './Select'
+export * from '../Select'
